+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [1.30.0] - 2023-02-03
+### Added
+- Update role admin API [#516](https://github.com/rokwire/core-building-block/issues/516)
+- Add admin application config endpoints [#633](https://github.com/rokwire/core-building-block/issues/633)
+### Changed
+- Upgrade dependencies [#624](https://github.com/rokwire/core-building-block/issues/624)
+### Fixed
+- Fix system account initialization [#594](https://github.com/rokwire/core-building-block/issues/594)
 
 ## [1.29.0] - 2022-11-16
 ### Added
@@ -356,7 +364,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Users authentication polish [#155](https://github.com/rokwire/core-building-block/issues/155)
 - Optimise the Mongo DB collections indexes usage [#146](https://github.com/rokwire/core-building-block/issues/146)
 
-[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.29.0...HEAD
+[Unreleased]: https://github.com/rokwire/core-building-block/compare/v1.30.0...HEAD
+[1.30.0]: https://github.com/rokwire/core-building-block/compare/v1.29.0...v1.30.0
 [1.29.0]: https://github.com/rokwire/core-building-block/compare/v1.28.0...v1.29.0
 [1.28.0]: https://github.com/rokwire/core-building-block/compare/v1.27.0...v1.28.0
 [1.27.0]: https://github.com/rokwire/core-building-block/compare/v1.26.0...v1.27.0